fix: address top style warnings
