+++ Dto/SocialMedia.cs
@@ -1,0 +1,17 @@
+// <copyright file="SocialMedia.cs" company="PlaceholderCompany">
+// Copyright (c) PlaceholderCompany. All rights reserved.
+// </copyright>
+
+namespace Dto;
+
+/// <summary>
+/// Social media links.
+/// </summary>
+public class SocialMedia
+{
+  public required string Facebook { get; set; }
+
+  public required string Twitter { get; set; }
+
+  public required string Instagram { get; set; }
+}
